Simplify apps model
The current code iterates over each service group separately and adds its apps. This leads to duplicate entries when an app is in multiple groups.

Since we are not actually interested in the group info, only in a flat list of all apps with some filters, we can simplify this.