mmplugin: set all connections to not autoconnect on setMobileDataEnabled(false) (#182)
When disabling mobile data, instead of setting the autoconnect flag to
true for the current active connection, set it to false for all
connections without exception. When reenabling mobile data, instead of
looking for a connection with autoconnect set, look for the connection
with the latest last use timestamp, set that one to autoconnect, and
also connect it immediately.

There should not be any connection with autoconnect enabled if mobile
data is disabled, because that makes NetworkManager reconnect it after a
reboot or a modem restart. The autoconnect state on the modem is
apparently lost on reboot.

TODO: We also need to change SignalIndicator::mobileDataEnabled(), or
it will still come out as true after the reboot.

Fixes #182.