panel: Overlay over lockscreen
Overlay the shell's status panel and quicksettings panel over the lockscreen, instead of rendering a second copy in the lockscreen theme. This will allow us to improve the lockscreen loading speed.

Key changes:

Overlay quicksettings window and the status bar over the lockscreen when it is shown
Refactor the top panel's showing logic to be cleaner (as it supports various overlay modes over fullscreen apps already)
Implement lockscreen support to the status bar and quicksettings panel in the to panel
Forward quicksettings panel requests for "unlock" over DBus to the lockscreen
Add "raiselockscreen" QML plugin to easily request a window to be raised over the lockscreen

Notes:

Now that we are sharing the quicksettings panel from the shell, notifications that are already there will be shown on the lockscreen (compared to right now, where only new notifications would be shown)