homescreen/folio: fix folder drop position misalignment and outside folder view check
This fixes an issue with the folder app icon drop position being misaligned. Noticeably causing issues when in gesture only mode.
Also, this fixes an issue with the outsideView calculation not accounting for folder size margins. Causing issues with some app still being interactive when outside the current folder page.