[info kcm]: add vendorinfo card
This change adds a card to the Information KCM which can be used to
display vendor-specific information (think of product info, serial or
model number, support info, etc.).

The information is read from a json file in /etc/vendorinfo.json, format
as follows:

'''
{
    "Title" : "Vendor Information",
    "Content" : [
        {
            "Key": "Model Number",
            "Value": "24ABC-13N4"
        },
        {
            "Key": "Serial Number",
            "Value": "778899223344"
        },
        {
            "Key": "Support Phone",
            "Value": "+31 6 48370928"
        }
    ]
}
'''

Title is the card header, each Key/Value block adds a TextDelegate. This
block is only displayed if the file actually exists and has a Title set.

Signed-off-by: Sebastian Kügler <sebas@kde.org>