new window management script
This changeset implements more fine-grained window management using a
javascript kwin script and the placement callback in this MR:
https://invent.kde.org/plasma/kwin/-/merge_requests/8079

Eventually, this should supersede the convergentwindows script and some
of our kwinrc config magic.

There's a change to plasma-mobile-envmanager here which makes the
plasmashell load and unload the new mobilewindows kwinscript when
switching between "mobile" and convergenceMode.

Setting this to draft since we first need to pin down the kwin
internals.

Closing this one.

Signed-off-by: Sebastian Kügler <sebas@kde.org>