panels: Add support for defining device specific panel tweaks
This adds support for specifying options needed to deal with phone
display panel pecularities (ex. screen curves, notches, punch holes)

This is implemented as settings in ~/.config/plasmamobilerc, which can
set panel heights, paddings, and center spacings to duck display
cutouts. The pixel values are scaling independent, and so are not
affected when the display scaling is changed.

This is then exposed over DBus, so that components from outside of
plasmashell (ex. KWin) can access it easily without needing to connect to
kscreen themselves. Each screen is exposed as a single object.

Currently support is only added in the status bar and the navigation
panel.

TODO:
- Ship premade device options
- Rework the config specification to be more generic, and just specify
  "safe zones" (ex. top -> left padding, panel height, right padding,
center spacing, left -> ...)
- Figure out how to obtain the screen IDs in plasmoids and kwin
- Figure out how this interacts with the current status bar scaling
  option
- Add good documentation for manually tweaking these options