panels: Add support for defining device specific panel tweaks
This adds support for specifying options needed to deal with phone
display panel pecularities (ex. screen curves, notches, punch holes)

This is implemented as settings in ~/.config/plasmamobilerc, which can
set panel heights, paddings, and center spacings to duck display
cutouts. The pixel values are scaling independent, and so are not
affected when the display scaling is changed.

Currently support is only added in the status bar and the navigation
panel.

TODO:
- Ship premade device options
- Move panelsettings to mobileshellstate, so that kscreen API calls are
  done from plasmashell (from KWin, we will get panel info over DBus)
- Rework the config specification to be more generic, and just specify
  "safe zones" (ex. top -> left padding, panel height, right padding,
center spacing, left -> ...)
- Figure out how to do per-screen (should expose over each
  screen over DBus), and how to obtain the screen IDs
- Figure out how this interacts with the current status bar scaling
  option
- Add good documentation for manually tweaking these options